requirements.txt
mlxtend==0.17.3
pandas==1.1.2
numpy==1.19.2
ipython==7.18.1
matplotlib==3.3.2
ipykernel==5.3.4
flake8==3.8.4
xlrd==1.2.0
networkx==2.5
seaborn==0.11.0
plotly==4.11.0
dash==1.16.3
Flask-SQLAlchemy==2.4.4
python-dotenv==0.14.0
gunicorn==20.0.4
PyMySQL==0.10.1
sentry-sdk[flask]==1.0.0
celery==5.0.5
redis==3.5.3
supervisor==4.2.1
gevent==21.1.2
